Use standard size_t type in the casts for length-parameter of memcpy, memmove and bzero.  When the library was written (1989), none of those had been standardized, and the source-code used "(int)" casts to help with K&R compilers.  The cleanup done in the previous update used binary-compare to validate, which does not work for these because the compiler is recording the cast's effect.
This change reduces the number of gcc warnings from 163 to 128.

+++ include/X11/IntrinsicI.h
@@ -110,14 +110,14 @@ SOFTWARE.
 
 #define XtMemmove(dst, src, size)	\
     if ((char *)(dst) != (char *)(src)) {		    \
-	(void) memcpy((char *) (dst), (char *) (src), (int) (size)); \
+	(void) memcpy((char *) (dst), (char *) (src), (size_t) (size)); \
     }
 
 #define XtBZero(dst, size) 	\
-	bzero((char *) (dst), (int) (size))
+	bzero((char *) (dst), (size_t) (size))
 
 #define XtMemcmp(b1, b2, size) 		\
-	memcmp((char *) (b1), (char *) (b2), (int) (size))
+	memcmp((char *) (b1), (char *) (b2), (size_t) (size))
 
 
 /****************************************************************
